pyproject.toml
[project]
name = "fastapi-crud-ecommerce"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafael Souza",email = "rafasansou1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
